style: move EmptyStatus to ProfileView dir
